requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-openai
langchain_community
langchain-chroma
chromadb
pypdf
python-dotenv
tiktoken
python-multipart